requirements.txt
alembic==1.16.5
anyio==4.11.0
certifi==2025.8.3
deepdiff==8.6.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.3.3
orderly-set==5.5.0
orjson==3.11.3
pandas==2.3.2
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
tenacity==9.1.2
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
